refactor: organize modules into domain packages
