+++ src/engine/processing_engine.py
@@ -18,7 +18,7 @@ from engine.service_execution import (
     ServiceExecution,
 )
 from engine.service_runtime import ServiceRuntime
-from loader import (
+from io_utils.loader import (
     configure_mapping_data_dir,
     configure_prompt_dir,
     load_evolution_prompt,
