+++ tests/test_e2e_cli_generate.py
@@ -72,11 +72,11 @@ dummy_loader = types.SimpleNamespace(
     load_prompt_text=lambda *a, **k: "",
     MAPPING_DATA_DIR=Path("data"),
 )
-sys.modules.setdefault("loader", cast(types.ModuleType, dummy_loader))
+sys.modules.setdefault("io_utils.loader", cast(types.ModuleType, dummy_loader))
 
 # Stub mapping and telemetry modules to satisfy CLI imports.
 sys.modules.setdefault(
-    "mapping",
+    "core.mapping",
     cast(
         types.ModuleType,
         types.SimpleNamespace(
@@ -87,7 +87,7 @@ sys.modules.setdefault(
     ),
 )
 sys.modules.setdefault(
-    "telemetry",
+    "observability.telemetry",
     cast(
         types.ModuleType,
         types.SimpleNamespace(
