+++ tests/test_e2e_cli_mapping.py
@@ -8,7 +8,7 @@ from pathlib import Path
 from types import SimpleNamespace
 
 import cli
-import mapping
+from core import mapping
 from models import Contribution, MappingSet
