+++ tests/test_e2e_cli_reverse.py
@@ -11,9 +11,8 @@ from typing import cast
 from pydantic_core import to_json
 
 import cli
-import loader
-import mapping
-import telemetry
+from core import mapping
+from io_utils import loader
 from models import (
     FeatureMappingRef,
     MappingFeatureGroup,
@@ -25,6 +24,7 @@ from models import (
     ServiceInput,
     ServiceMeta,
 )
+from observability import telemetry
 
 dummy_logfire = types.SimpleNamespace(
     metric_counter=lambda name: types.SimpleNamespace(add=lambda *a, **k: None),
